requirements.txt
streamlit
openai-whisper
fpdf
torch
torchaudio